imtixonYukjlandi

+++ src/global/coments/coments.service.ts
@@ -7,15 +7,17 @@ import { CustomErrorService } from 'src/core/custom-error/custom-error.service';
 @Injectable()
 export class ComentsService {
   constructor(private prisma: PrismaService, private customErrors: CustomErrorService) { }
-  async create(createComentDto: CreateComentDto, user_id: string) {
+  async create(createComentDto: CreateComentDto, user_id: string, slug: string) {
+    const checkUser = await this.prisma.permission.findUnique({ where: { userId: user_id } })
+    if (!checkUser || checkUser.can_add === false) throw new NotFoundException("user's permission not found or you don't alowed can_add!")
     const findUser = await this.customErrors.findByUserId(user_id)
-    const findRating = await this.prisma.movie.findFirst({ where: { rating: createComentDto.rating } })
-    if (!findRating) throw new NotFoundException("this movies's not found rating!")
+    const findSlug = await this.prisma.movie.findFirst({ where: { slug } })
+    if (!findSlug) throw new NotFoundException("this movies's not found slug!")
 
     const result = await this.prisma.review.create({
       data: {
-        movieId: findRating.id,
-        rating: findRating.rating,
+        movieId: findSlug.id,
+        rating: findSlug.rating,
         comment: createComentDto.comment,
         userId: user_id
       }
@@ -37,6 +39,8 @@ export class ComentsService {
     }
   }
   async remove(user_id: string, review_id: string, movie_id: string) {
+    const checkUser = await this.prisma.permission.findUnique({ where: { userId: user_id } })
+    if (!checkUser || checkUser.can_delete === false) throw new NotFoundException("user's permission not found or you don't alowed can_delete!")
     await this.customErrors.findByUserId(user_id)
     await this.customErrors.findMovieById(movie_id)
     const findReview = await this.prisma.review.findUnique({ where: { id: review_id } })
@@ -73,7 +77,9 @@ export class ComentsService {
       data: result
     }
   }
-  async deleteAllReviews(movie_id: string) {
+  async deleteAllReviews(movie_id: string, user_id: string) {
+    const checkUser = await this.prisma.permission.findUnique({ where: { userId: user_id } })
+    if (!checkUser || checkUser.can_delete === false) throw new NotFoundException("user's permission not found or you don't alowed can_delete!")
     await this.prisma.review.deleteMany({ where: { movieId: movie_id } })
     return {
       success: true,
